pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet python course: project 3"
packages = [{include = "page_analyzer"}]
authors = ["Alina Li <alinali87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
